image/requirements.txt
Flask
requests
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
jinja2<3.1.0